A new configuration type: duration
Introduce a new configuration type to be parsed. The duration is
an ISO 8601 duration that is going to be used for DNSSEC key timings
such as key lifetimes, signature resign intervals and refresh
periods, etc.

A duration is stored as an array of 7 different time parts.
A duration can either be expressed in weeks, or in a combination of
the other datetime indicators.

Add several unit tests to ensure the correct value is parsed given
different string values.