Change the dns_cache_dump() to private static cache_dump() function
The dns_cache_dump() public function was used only internally when destroying
the cache and it checked the validity of the cache object when in fact the cache
object was no longer valid (.references == 0).  This commit changes the
dns_cache_dump() function into static cache_dump() function without the validity
check.