Introduce keymgr function in named
Add a keymgr function to named.  Currently this only checks if there
are any keys at all, and if not it creates DNSSEC keys according to
the dnssec-policy.

Remove 'allzsk' from 'dns_dnssec_updatekeys' as this was only a
hint for logging.

Add a new variable to 'struct dns_dnsseckey' to signal whether this
key is a zone-signing key (it is no longer true that ksk == !zsk).

Update 'dns_dnssec_findzonekeys' and 'dns_dnssec_findmatchingkeys'
to also read the key state file, if available.