Introduce keymgr inside named
If a dnssec-policy is set, 'named' will run a key manager on the
matching keys. This will do a couple of keys:

1. Create keys when needed (in case of rollover for example)
   according to the set policy.

2. Retire keys that are in excess of the policy.

3. Maintain key states according to "Flexible and Robust Key
   Rollover" [1]. After key manager ran, key files will be saved to
   disk.

   [1] https://matthijsmekking.nl/static/pdf/satin2012-Schaeffer.pdf

This commit comes with just a few tests to test key creation at
start up, following commits need to add more test cases.  It
extends the kasp system test to also verify key states.

There are small changes to the dnssec.c file (in addition to earlier
commits).  For this commit it is important to have a signal that
a key wants to be retired. This will determine the flow of the
key state machines (a key that wants to be active wants its records
in OMNIPRESENT state, and a key that wants to retire wants its
records in HIDDEN state).

Also make get_hints() (now dns_dnssec_get_hints()) public so that
we can use it in the key manager.

The dst code adds ways to write and read the new key states and
timing metadata. It updates the parsing code for private key files
to not parse the newly introduced metadata (these are for state
files only).

Finally for comparisons between keys and kasp key configuration,
update the dns_kasp_key_size() funtion to return unsigned int.