Test signature refresh
This commit adds two more tests and code changes in order to
successfully verify that expired and unfresh signatures will be
regenerated.

First of all, another refactor of the kasp system test is needed
such that KSK, ZSK1 and ZSK2 expected states can be stored (in an
array) prior to calling `check_zone`.

The tests make use of the earlier introduced dnssec-settime key
state options to set key states on keys that are omnipresent for a
while.

In order to make those tests pass we need to relax the rules when
automated dnssec applies:
- A dnssec-policy on a zone now means that we create a raw zone as
  if we were doing inline signing.
- Dynamic in the context of DNSSEC maintenance means either a zone
  that is dynamically updated, or has a dnssec-policy.