keygen/settime: Write out successor/predecessor
When creating a successor key, or calculating time for a successor
key, write out the successor and predecessor metadata to the
related files.

Also, for dnssec-settime, extend the setting of the key states with
timing metadata, so you can control when the state change occurred.

This commit sneaks in a bugfix: the order of timingtags in
dst_api.c was wrong.

This commit also slightly updates the 'dst_key_is_active()'
function, such that key states trump the `inactive` timing metadata.