Convert the magic numbers to standard atomics
The magic number access has been done in quite frivolous manner
because it's usually protected by the reference counting.

Unfortunately, it prevents the instance being used only when
any user of the instance correctly attaches / detaches to the
object.

For a long term, we should consider the instance of the object valid
only if:

* it's magic number is valid
* the reference counter is not 0

Both values needs to be accessed in atomic manner in a struct, so we
have assurance, that we are using expected object and it hasn't been
invalidated by a different thread.

This is the first part of the changes, the next step must be changing
the destroy functions to invalidate the magic as soon as possible.