Downgrade the MSVC WarningLevel to Level2
Our vcxproj files set the WarningLevel to Level3, which is too verbose
for a code that needs to be portable.  That basically leads to ignoring
all the errors that MSVC produces.  This commits downgrades the
WarningLevel to Level2, so we reduce the set of the warnings to a level
where we can use /WX (MSVC equivalent of cc -Werror).