use asynchronous handlers for rndc where necesary
some rndc commands call functions which need to run exclusively, and
which are not only run by rndc but also under the ISC taskmgr. in order
for isc_task_beginexclusive() to work correctly, these functions must
be run in asynchronous task events, so they have been modified to do so.
the affected commands are:

- rndc reconfig
- rndc reload (without a zone name)
- rndc scan

one side effect of this change is that "rndc reconfig" can no longer
return a failure status directly to the caller; faliure to load
configuration is only visible in the log.