implement isc_nm_beginexclusive() and isc_nm_endexclusive()
these functions mirror isc_task_beginexclusive() and
isc_task_endexclusive(). when entering exclusive mode, every
netmgr worker is stopped except the one that's currently running,
and the taskmgr is paused, so that the current netmgr event will be
the only thing running in the system. exclusive mode must be
relinquished before the event handler returns.

NOTE: this can deadlock if the isc_nm_beginexclusive() is
called while another thread is waiting for isc_nm_pause() to
finish. this needs further work.