restore the call to testsummary.sh
after running make check, call the testsummary.sh script.

when running from the automake-generated makefile, use the -s flag;
this suppresses printing of the counts of passed, failed, and skipped
tests. this information would be redundant, since the standard makefile
output already includes such a summary.  the important part is to report
the *names* of failed tests and tests with core dumps or assertion
failures.