Fix abort() when stopping named
Due to lack of synchronization, whenever named was being requested to
stop using rndc, controlconf.c module was trying to access an already
released pointer through named_g_server->interfacemgr.

It was added an atomic bool variable to the named_g_server instance,
which is used to check wheter named is in process of shutting down,
and if so, avoid accessing named_g_server->interfacemgr which may have
been destroyed already.